# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ML Trading System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ModelMetrics {
  accuracy: number;
  auc: number;
  sharpe?: number;
  win_rate_pct?: number;
  n_trades?: number;
}

export type SummaryResponseModels = { [key: string]: ModelMetrics };

export interface TomorrowPrediction {
  direction: string;
  confidence: number;
  strength: string;
}

export type SummaryResponseTomorrow = { [key: string]: TomorrowPrediction };

export interface TickerRankingEntry {
  rank: number;
  probability: number;
  direction: string;
  strength: string;
}

export type SummaryResponseTickerRanking = { [key: string]: TickerRankingEntry };

export type SummaryResponseDiagnostics = { [key: string]: unknown };

export interface SummaryResponse {
  status: string;
  run_time?: string;
  ticker?: string;
  best_model?: string;
  n_features?: number;
  prediction_horizon?: number;
  forecast_label?: string;
  models?: SummaryResponseModels;
  tomorrow?: SummaryResponseTomorrow;
  ticker_ranking?: SummaryResponseTickerRanking;
  diagnostics?: SummaryResponseDiagnostics;
}

export type ConfigResponseConfig = { [key: string]: unknown };

export interface ConfigResponse {
  config: ConfigResponseConfig;
  raw_yaml: string;
}

export interface SaveConfigRequest {
  raw_yaml: string;
}

export interface SaveConfigResponse {
  success: boolean;
  message: string;
}

export interface ChartListResponse {
  charts: string[];
}

export interface LogListResponse {
  logs: string[];
}

export interface LogContentResponse {
  filename: string;
  content: string;
}

export type PredictionHistoryResponseRowsItem = { [key: string]: unknown };

export interface PredictionHistoryResponse {
  rows: PredictionHistoryResponseRowsItem[];
  count: number;
}

export interface RunStartResponse {
  success: boolean;
  message: string;
  already_running?: boolean;
}

export type RunStatusResponseStatus =
  (typeof RunStatusResponseStatus)[keyof typeof RunStatusResponseStatus];

export const RunStatusResponseStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
  idle: "idle",
} as const;

export interface RunStatusResponse {
  status: RunStatusResponseStatus;
  return_code?: number | null;
  current_step?: number;
  log_tail?: string;
}

export type QuickPredictResponsePredictions = { [key: string]: unknown };

export interface QuickPredictResponse {
  success: boolean;
  ticker: string;
  predictions?: QuickPredictResponsePredictions;
  error?: string | null;
}