# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ML Trading System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get config.yaml contents
      responses:
        "200":
          description: Config as key-value object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
    post:
      operationId: saveConfig
      tags: [config]
      summary: Save config.yaml
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveConfigResponse"
...
  /logs/{filename}:
    get:
      operationId: getLog
      tags: [logs]
      summary: Get log file contents
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log file content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogContentResponse"
...
  /predict/{ticker}:
    get:
      operationId: quickPredict
      tags: [pipeline]
      summary: Run quick prediction for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickPredictResponse"
components:
  schemas:
...
    ConfigResponse:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
        raw_yaml:
          type: string
      required:
        - config
        - raw_yaml
    SaveConfigRequest:
      type: object
      properties:
        raw_yaml:
          type: string
      required:
        - raw_yaml
    SaveConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    LogContentResponse:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
      required:
        - filename
        - content
...
    QuickPredictResponse:
      type: object
      properties:
        success:
          type: boolean
        ticker:
          type: string
        predictions:
          type: object
          additionalProperties: true
        error:
          type: string
          nullable: true
      required:
        - success
        - ticker