# packages/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ML Trading System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get pipeline results summary
 */
export const GetSummaryResponse = zod.object({
  status: zod.string(),
  run_time: zod.string().optional(),
  ticker: zod.string().optional(),
  best_model: zod.string().optional(),
  n_features: zod.number().optional(),
  prediction_horizon: zod.number().optional(),
  forecast_label: zod.string().optional(),
  models: zod
    .record(
      zod.string(),
      zod.object({
        accuracy: zod.number(),
        auc: zod.number(),
        sharpe: zod.number().optional(),
        win_rate_pct: zod.number().optional(),
        n_trades: zod.number().optional(),
      }),
    )
    .optional(),
  tomorrow: zod
    .record(
      zod.string(),
      zod.object({
        direction: zod.string(),
        confidence: zod.number(),
        strength: zod.string(),
      }),
    )
    .optional(),
  ticker_ranking: zod
    .record(
      zod.string(),
      zod.object({
        rank: zod.number(),
        probability: zod.number(),
        direction: zod.string(),
        strength: zod.string(),
      }),
    )
    .optional(),
  diagnostics: zod.record(zod.string(), zod.unknown()).optional(),
});

/**
 * @summary Get config.yaml contents
 */
export const GetConfigResponse = zod.object({
  config: zod.record(zod.string(), zod.unknown()),
  raw_yaml: zod.string(),
});

/**
 * @summary Save config.yaml
 */
export const SaveConfigBody = zod.object({
  raw_yaml: zod.string(),
});

export const SaveConfigResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary List available chart PNG files
 */
export const ListChartsResponse = zod.object({
  charts: zod.array(zod.string()),
});

/**
 * @summary List available log files
 */
export const ListLogsResponse = zod.object({
  logs: zod.array(zod.string()),
});

/**
 * @summary Get log file contents
 */
export const GetLogParams = zod.object({
  filename: zod.coerce.string(),
});

export const GetLogResponse = zod.object({
  filename: zod.string(),
  content: zod.string(),
});

/**
 * @summary Get prediction history from CSV
 */
export const GetPredictionHistoryResponse = zod.object({
  rows: zod.array(zod.record(zod.string(), zod.unknown())),
  count: zod.number(),
});

/**
 * @summary Start the ML pipeline (python main.py)
 */
export const StartRunResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  already_running: zod.boolean().optional(),
});

/**
 * @summary Poll run status
 */
export const GetRunStatusResponse = zod.object({
  status: zod.enum(["running", "completed", "failed", "idle"]),
  return_code: zod.number().nullish(),
  log_tail: zod.string().optional(),
});

/**
 * @summary Run quick prediction for a ticker
 */
export const QuickPredictParams = zod.object({
  ticker: zod.coerce.string(),
});

export const QuickPredictResponse = zod.object({
  success: zod.boolean(),
  ticker: zod.string(),
  predictions: zod.record(zod.string(), zod.unknown()).optional(),
  error: zod.string().nullish(),
});