# packages/api-zod/src/generated/types/logContentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ML Trading System API
 * OpenAPI spec version: 0.1.0
 */

export interface LogContentResponse {
  filename: string;
  content: string;
}