# packages/api-zod/src/generated/types/saveConfigResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ML Trading System API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveConfigResponse {
  success: boolean;
  message: string;
}